All MsDataHub datasets

The MsDataHub package provides example mass spectrometry data, peptide spectrum matches or quantitative data from proteomics and metabolomics experiments.

The MsDataHub() function returns a data.frame with all the annotated datasets provided in the package. For details on these individual datasets, refer to their respective manual pages.

See the vignette and the respective manuals pages for more details about the package and the data themselves.

MsDataHub()

Value

A data.frame describing the data available in MsDataHub.

Author

Laurent Gatto

Examples

MsDataHub()
#>                                                                     Title
#> 1                                                                ko15.CDF
#> 2                                                cptac_a_b_c_peptides.txt
#> 3                                                  cptac_a_b_peptides.txt
#> 4                                                      cptac_peptides.txt
#> 5     TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.20141210.mzid
#> 6  TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.20141210.mzML.gz
#> 7                                       X20171016_POOL_POS_1_105.134.mzML
#> 8                                       X20171016_POOL_POS_3_105.134.mzML
#> 9                                                       PestMix1_DDA.mzML
#> 10                                                    PestMix1_SWATH.mzML
#> 11                                                    benchmarkingDIA.tsv
#> 12                                           Report.Derks2022.plexDIA.tsv
#> 13                                                 Ai2025_aCMs_report.tsv
#> 14                                                 Ai2025_iCMs_report.tsv
#> 15                                                         crap_gpm.fasta
#> 16                                                         crap_ccp.fasta
#> 17                                                 crap_maxquant.fasta.gz
#>                                                                                                                                Description
#> 1                                                                    Raw metabolomics MS file in netCDF format. See ?ko15.CDF for details.
#> 2                          Conditions A, B and C of the CPTAC quantitative proteomics data (tab-delimited format). See ?cptac for details.
#> 3                             Conditions A and B of the CPTAC quantitative proteomics data (tab-delimited format). See ?cptac for details.
#> 4                                                       CPTAC quantitative proteomics data (tab-delimited format). See ?cptac for details.
#> 5                                                      Peptide spectrum matches from the PDX000001 experiment. See ?PXD000001 for details.
#> 6                                                   Raw MS data from the PDX000001 experiment, in mzML format. See ?PXD000001 for details.
#> 7                                                    AB Sciex LC-MS data file (injection index 1), in mzML format. See ?sciex for details.
#> 8                                                   AB Sciex LC-MS data file (injection index 19), in mzML format. See ?sciex for details.
#> 9                                                                     Triple TOF DDA raw data, in mzML format. See ?TripleTOF for details.
#> 10                                                                  Triple TOF SWATH raw data, in mzML format. See ?TripleTOF for details.
#> 11                                                                                                  Output of DIA-NN software (report.tsv)
#> 12                                                          Derk et al. (2022) single-cell proteomics plexDIA results (DIA-NN report.tsv).
#> 13 Ai et al. (2025) Single Cell Proteomics Reveals Specific Cellular Subtypes in Cardiomyocytes Derived from Human iPSCs and Adult Hearts.
#> 14 Ai et al. (2025) Single Cell Proteomics Reveals Specific Cellular Subtypes in Cardiomyocytes Derived from Human iPSCs and Adult Hearts.
#> 15                                                common Repository of Adventitious Proteins (cRAP) from the Global Proteome Machine (GPM)
#> 16                                        common Repository of Adventitious Proteins (cRAP) from the Cambridge Centre for Proteomics (CCP)
#> 17                                                                                                         MaxQuant's contaminant database
#>    BiocVersion Genome SourceType
#> 1         3.17     NA        CDF
#> 2         3.17     NA        TXT
#> 3         3.17     NA        TXT
#> 4         3.17     NA        TXT
#> 5         3.17     NA       mzid
#> 6         3.17     NA       mzML
#> 7         3.17     NA       mzML
#> 8         3.17     NA       mzML
#> 9         3.17     NA       mzML
#> 10        3.17     NA       mzML
#> 11        3.17     NA        TSV
#> 12        3.19     NA        TSV
#> 13        3.21     NA        TSV
#> 14        3.21     NA        TSV
#> 15        3.21     NA      FASTA
#> 16        3.21     NA      FASTA
#> 17        3.21     NA      FASTA
#>                                                                                                                  SourceUrl
#> 1                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 2                                                 https://uclouvain-cbio.github.io/WSBIM2122/data/cptac_a_b_c_peptides.txt
#> 3                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 4         https://raw.githubusercontent.com/statOmics/PDA/data/quantification/fullCptacDatasSetNotForTutorial/peptides.txt
#> 5                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 6                                                        https://ftp.pride.ebi.ac.uk/pride/data/archive/2012/03/PXD000001/
#> 7                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 8                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 9                                                  https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 10                                                 https://bioconductor.org/packages/3.16/data/experiment/html/msdata.html
#> 11                                                                                      https://zenodo.org/records/8063173
#> 12                                                https://drive.google.com/drive/folders/1pUC2zgXKtKYn22mlor0lmUDK0frgwL_-
#> 13  ftp://massive.ucsd.edu/v07/MSV000094438/updates/2024-10-23_bineka_5deafa61/other/Supplementary%20Files/aCMs/report.tsv
#> 14 ftp://massive.ucsd.edu/v07/MSV000094438/updates/2024-10-23_bineka_5deafa61/other/Supplementary%20Files/iCMs/report1.tsv
#> 15                                                                https://zenodo.org/records/15115102/files/crap_gpm.fasta
#> 16                                                                https://zenodo.org/records/15115102/files/crap_ccp.fasta
#> 17                                                        https://zenodo.org/records/15115102/files/crap_maxquant.fasta.gz
#>    SourceVersion                  Species TaxonomyId Coordinate_1_based
#> 1              1             Mus musculus      10090                 NA
#> 2              1 Saccharomyces cerevisiae       4932                 NA
#> 3              1 Saccharomyces cerevisiae       4932                 NA
#> 4              1 Saccharomyces cerevisiae       4932                 NA
#> 5              1       Erwinia carotovora        554                 NA
#> 6              1       Erwinia carotovora        554                 NA
#> 7              1             Homo sapiens       9606                 NA
#> 8              1             Homo sapiens       9606                 NA
#> 9              1                                  NA                 NA
#> 10             1                                  NA                 NA
#> 11             1             Homo sapiens       9606                 NA
#> 12             1             Homo sapiens       9606                 NA
#> 13             1             Homo sapiens       9606                 NA
#> 14             1             Homo sapiens       9606                 NA
#> 15             1                                  NA                 NA
#> 16             1                                  NA                 NA
#> 17             1                                  NA                 NA
#>    DataProvider                                 Maintainer  RDataClass
#> 1            NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 2            NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 3            NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 4            NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 5            NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 6            NA Laurent Gatto <laurent.gatto@uclouvain.be>         PSM
#> 7            NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 8            NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 9            NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 10           NA Laurent Gatto <laurent.gatto@uclouvain.be>     Spectra
#> 11           NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 12           NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 13           NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 14           NA Laurent Gatto <laurent.gatto@uclouvain.be>  data.frame
#> 15           NA Laurent Gatto <laurent.gatto@uclouvain.be> AAStringSet
#> 16           NA Laurent Gatto <laurent.gatto@uclouvain.be> AAStringSet
#> 17           NA Laurent Gatto <laurent.gatto@uclouvain.be> AAStringSet
#>    DispatchClass         Location_Prefix
#> 1       FilePath                        
#> 2       FilePath                        
#> 3       FilePath                        
#> 4       FilePath                        
#> 5       FilePath                        
#> 6       FilePath                        
#> 7       FilePath                        
#> 8       FilePath                        
#> 9       FilePath                        
#> 10      FilePath                        
#> 11      FilePath     https://zenodo.org/
#> 12      FilePath     https://zenodo.org/
#> 13      FilePath ftp://massive.ucsd.edu/
#> 14      FilePath ftp://massive.ucsd.edu/
#> 15      FilePath     https://zenodo.org/
#> 16      FilePath     https://zenodo.org/
#> 17      FilePath     https://zenodo.org/
#>                                                                                           RDataPath
#> 1                                                                            MsDataHub/cdf/ko15.CDF
#> 2                                                          MsDataHub/cptac/cptac_a_b_c_peptides.txt
#> 3                                                            MsDataHub/cptac/cptac_a_b_peptides.txt
#> 4                                                                MsDataHub/cptac/cptac_peptides.txt
#> 5           MsDataHub/PXD000001/TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid
#> 6        MsDataHub/PXD000001/TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML.gz
#> 7                                                  MsDataHub/sciex/20171016_POOL_POS_1_105-134.mzML
#> 8                                                  MsDataHub/sciex/20171016_POOL_POS_3_105-134.mzML
#> 9                                                       MsDataHub/TripleTOF-SWATH/PestMix1_DDA.mzML
#> 10                                                    MsDataHub/TripleTOF-SWATH/PestMix1_SWATH.mzML
#> 11                                                         record/8063173/files/benchmarkingDIA.tsv
#> 12                                              records/10938597/files/Report-Derks2022-plexDIA.tsv
#> 13  v07/MSV000094438/updates/2024-10-23_bineka_5deafa61/other/Supplementary%20Files/aCMs/report.tsv
#> 14 v07/MSV000094438/updates/2024-10-23_bineka_5deafa61/other/Supplementary%20Files/iCMs/report1.tsv
#> 15                                                            records/15115102/files/crap_gpm.fasta
#> 16                                                            records/15115102/files/crap_ccp.fasta
#> 17                                                    records/15115102/files/crap_maxquant.fasta.gz
#>    Tags
#> 1    NA
#> 2    NA
#> 3    NA
#> 4    NA
#> 5    NA
#> 6    NA
#> 7    NA
#> 8    NA
#> 9    NA
#> 10   NA
#> 11   NA
#> 12   NA
#> 13   NA
#> 14   NA
#> 15   NA
#> 16   NA
#> 17   NA