Triple TOF SWATH Data

These files represent data from reverse-phased LC-MS/MS runs on the Agilent Pesticide mix obtained from a Sciex 6600 Triple ToF operated either in Sequential Window Acquisition of all THeoretical mass spectra (SWATH) or Data Dependent Acquisition (DDA) acquisition mode.

The data files are:

PestMix1_DDA.mzML: mzML file with MS1 and MS2 spectra from the Agilent Pesticide Mix acquired in DDA mode.

PestMix1_SWATH.mzML: mzML file with MS1 and MS2 spectra from the Agilent Pesticide Mix acquired in SWATH mode.

Author

Micheal Witting, Johannes Rainer