Misc DDA and DIA datasets

Various DDA and DIA experiments that were processed with different software to illustrate the QFeatures::readQFeatures() function:

vanPuyvelde_2022_LFQ_*: DDA and DIA data from van Puyvelde et al. (2022). LFQ benchmarking experiment.

Christoforou_2016_TMT_DDA_*: data from Christoforou et al. (2016). hyperLOPIT spatial proteomics experiment.

Derks_2022_plex_DIA_DIANN_report_subset.tsv: subset from Derks et al. (2022) plexDIA data.

Source files are available on Zenodo: https://zenodo.org/records/19137577

References

Van Puyvelde, B., Daled, S., Willems, S. et al. A comprehensive LFQ benchmark dataset on modern day acquisition strategies in proteomics. Sci Data 9, 126 (2022). https://doi.org/10.1038/s41597-022-01216-6.

Christoforou, A., Mulvey, C., Breckels, L. et al. A draft map of the mouse pluripotent stem cell spatial proteome. Nat Commun 7, 9992 (2016). https://doi.org/10.1038/ncomms9992

Derks, J., Leduc, A., Wallmann, G. et al. Increasing the throughput of sensitive proteomics by plexDIA. Nat Biotechnol 41, 50–59 (2023). https://doi.org/10.1038/s41587-022-01389-w

Author

Kristina Gomoryova and Laurent Gatto